+++ src/configuration.ts
@@ -18,6 +18,15 @@
 import * as fs from "fs";
 import * as path from "path";
 import * as findup from "findup-sync";
+import * as pathIsAbsolute from "path-is-absolute";
+
+import {arrayify, objectify} from "./utils";
+
+export interface IConfigurationFile {
+    extends?: string | string[];
+    rulesDirectory?: string | string[];
+    rules?: any;
+}
 
 export const CONFIG_FILENAME = "tslint.json";
 export const DEFAULT_CONFIG = {
@@ -53,7 +62,7 @@ export const DEFAULT_CONFIG = {
 };
 
 const PACKAGE_DEPRECATION_MSG = "Configuration of TSLint via package.json has been deprecated, "
-   + "please start using a tslint.json file instead (http://palantir.github.io/tslint/usage/tslint-json/).";
+    + "please start using a tslint.json file instead (http://palantir.github.io/tslint/usage/tslint-json/).";
 
 /**
  * Searches for a TSLint configuration and returns the data from the config.
@@ -62,7 +71,7 @@ const PACKAGE_DEPRECATION_MSG = "Configuration of TSLint via package.json has be
  * of the search for a configuration.
  * @returns A TSLint configuration object
  */
-export function findConfiguration(configFile: string, inputFilePath: string): any {
+export function findConfiguration(configFile: string, inputFilePath: string): IConfigurationFile {
     const configPath = findConfigurationPath(configFile, inputFilePath);
     return loadConfigurationFromPath(configPath);
 }
@@ -74,7 +83,7 @@ export function findConfiguration(configFile: string, inputFilePath: string): an
  * the location of the config file is not known and you want to search for one.
  * @param inputFilePath A path to the current file being linted. This is the starting location
  * of the search for a configuration.
- * @returns A path to a tslint.json file, a path to a package.json file with a tslintConfig field
+ * @returns An absolute path to a tslint.json file, a path to a package.json file with a tslintConfig field
  * or undefined if neither can be found.
  */
 export function findConfigurationPath(suppliedConfigFilePath: string, inputFilePath: string) {
@@ -82,19 +91,19 @@ export function findConfigurationPath(suppliedConfigFilePath: string, inputFileP
         if (!fs.existsSync(suppliedConfigFilePath)) {
             throw new Error(`Could not find config file at: ${path.resolve(suppliedConfigFilePath)}`);
         } else {
-            return suppliedConfigFilePath;
+            return path.resolve(suppliedConfigFilePath);
         }
     } else {
         // search for tslint.json from input file location
         let configFilePath = findup(CONFIG_FILENAME, { cwd: inputFilePath, nocase: true });
         if (configFilePath != null && fs.existsSync(configFilePath)) {
-            return configFilePath;
+            return path.resolve(configFilePath);
         }
 
         // search for package.json with tslintConfig property
         configFilePath = findup("package.json", { cwd: inputFilePath, nocase: true });
         if (configFilePath != null && require(configFilePath).tslintConfig != null) {
-            return configFilePath;
+            return path.resolve(configFilePath);
         }
 
         // search for tslint.json in home directory
@@ -102,7 +111,7 @@ export function findConfigurationPath(suppliedConfigFilePath: string, inputFileP
         if (homeDir != null) {
             configFilePath = path.join(homeDir, CONFIG_FILENAME);
             if (fs.existsSync(configFilePath)) {
-                return configFilePath;
+                return path.resolve(configFilePath);
             }
         }
 
@@ -112,20 +121,75 @@ export function findConfigurationPath(suppliedConfigFilePath: string, inputFileP
 }
 
 /**
- * @returns a configuration object for TSLint loaded form the file at configFilePath
+ * Used Node semantics to load a configuration file given configFilePath.
+ * For example:
+ * '/path/to/config' will be treated as an absolute path
+ * './path/to/config' will be treated as a relative path
+ * 'path/to/config' will attempt to load a to/config file inside a node module named path
+ * @returns a configuration object for TSLint loaded from the file at configFilePath
  */
-export function loadConfigurationFromPath(configFilePath: string) {
+export function loadConfigurationFromPath(configFilePath: string): IConfigurationFile {
     if (configFilePath == null) {
         return DEFAULT_CONFIG;
     } else if (path.basename(configFilePath) === "package.json") {
         console.warn(PACKAGE_DEPRECATION_MSG);
         return require(configFilePath).tslintConfig;
     } else {
-        let fileData = fs.readFileSync(configFilePath, "utf8");
-        // remove BOM if present
-        fileData = fileData.replace(/^\uFEFF/, "");
-        return JSON.parse(fileData);
+        const resolvedConfigFilePath = resolveConfigurationPath(configFilePath);
+        let configFile: IConfigurationFile = require(resolvedConfigFilePath);
+        const configFileDir = path.dirname(resolvedConfigFilePath);
+
+        configFile.rulesDirectory = getRulesDirectories(configFile.rulesDirectory, configFileDir);
+        configFile.extends = arrayify(configFile.extends);
+
+        for (const name of configFile.extends) {
+            const baseConfigFilePath = resolveConfigurationPath(name, configFileDir);
+            const baseConfigFile = loadConfigurationFromPath(baseConfigFilePath);
+            configFile = extendConfigurationFile(configFile, baseConfigFile);
+        }
+        return configFile;
     }
+}
+
+/**
+ * Resolve configuration file path
+ * @var relativeFilePath Relative path or package name (tslint-config-X) or package short name (X)
+ */
+function resolveConfigurationPath(relativeFilePath: string, relativeTo?: string) {
+    let resolvedPath: string;
+    if (pathIsAbsolute(relativeFilePath)) {
+        resolvedPath = relativeFilePath;
+    } else if (relativeFilePath.indexOf(".") === 0) {
+        resolvedPath = getRelativePath(relativeFilePath, relativeTo);
+    } else {
+        try {
+            resolvedPath = require.resolve(relativeFilePath);
+        } catch (err) {
+            throw new Error(`Invalid "extends" configuration value - could not require "${relativeFilePath}". ` +
+                "Review the Node lookup algorithm (https://nodejs.org/api/modules.html#modules_all_together) " +
+                "for the approximate method TSLint uses to find the referenced configuration file.");
+        }
+    }
+
+    return resolvedPath;
+}
+
+export function extendConfigurationFile(config: IConfigurationFile, baseConfig: IConfigurationFile): IConfigurationFile {
+    let combinedConfig: IConfigurationFile = {};
+
+    const baseRulesDirectory = arrayify(baseConfig.rulesDirectory);
+    const configRulesDirectory = arrayify(config.rulesDirectory);
+    combinedConfig.rulesDirectory = configRulesDirectory.concat(baseRulesDirectory);
+
+    combinedConfig.rules = {};
+    for (const name of Object.keys(objectify(baseConfig.rules))) {
+        combinedConfig.rules[name] = baseConfig.rules[name];
+    }
+    for (const name of Object.keys(objectify(config.rules))) {
+        combinedConfig.rules[name] = config.rules[name];
+    }
+
+    return combinedConfig;
 }
 
 function getHomeDir() {
@@ -159,15 +223,7 @@ export function getRelativePath(directory: string, relativeTo?: string): string 
  * @return An array of absolute paths to directories potentially containing rules
  */
 export function getRulesDirectories(directories: string | string[], relativeTo?: string): string[] {
-    let rulesDirectories: string[] = [];
-
-    if (directories != null) {
-        if (typeof directories === "string") {
-            rulesDirectories = [getRelativePath(directories, relativeTo)];
-        } else {
-            rulesDirectories = directories.map((dir) => getRelativePath(dir, relativeTo));
-        }
-    }
+    const rulesDirectories = arrayify(directories).map((dir) => getRelativePath(dir, relativeTo));
 
     for (const directory of rulesDirectories) {
         if (!fs.existsSync(directory)) {